01a.&03a.DataParsing/python_server/pyproject.toml
[tool.poetry]
name = "python_server"
version = "0.1.0"
description = ""
authors = ["asgerbirk <asger.birk@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
pyyaml = "^6.0.1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
